requirements.txt
mysql-connector-python~=8.0.27
inquirer~=2.7.0